+++ src/devices/lcd.py
@@ -1,7 +1,8 @@
 import time
 
-import constants
 import digitalio
+
+import constants
 
 
 class LCD:
