+++ src/devices/temp_probe.py
@@ -1,7 +1,8 @@
 import adafruit_max31865
 import busio
-import constants
 import digitalio
+
+import constants
 
 
 class Temp_Probe:
