+++ src/devices/temp_probe_mock.py
